+++ packages/components/tsup.config.ts
@@ -9,11 +9,9 @@ export default defineConfig({
   bundle: env === 'production',
   clean: true,
   dts: true,
-  format: ['esm', 'cjs'],
+  format: ['cjs'],
   external: ['react'],
-  splitting: true,
   minify: env === 'production',
-  skipNodeModulesBundle: true,
   watch: env === 'development',
   sourcemap: env === 'development',
   tsconfig: path.resolve(__dirname, './tsconfig.json')